fix(pctl): 修正游玩计时器布局和剩余时间读取
- 按 0x44 真机快照修正七天设置偏移并保留无关字段
- 将剩余时间按 u64 纳秒读取并转换为分钟
- 升级 capability backend 并记录探测前后及恢复后的 raw 快照
- 补充真实设备 fixture、边界回归测试和复测文档

+++ common/protocol/result_builder.h
@@ -13,7 +13,7 @@ typedef struct {
     int blocked_today;
     int unrestricted_today;
     bool remaining_available;
-    int remaining_minutes;
+    int64_t remaining_minutes;
     int play_timer_enabled;
     int restricted_now;
     bool bedtime_active;
